fix(release): ship gvproxy to Linux installs (native own-ip)
The Linux component set omitted gvproxy, so the native (DM2) own-IP
datapath was dead on every installed Linux host: minimald resolved the
switch only at the fixed /usr/lib/minimal/bin/gvproxy, which no install
provides. The release workflow already stages pin-verified
gvproxy-linux-{amd64,arm64} artifacts; they were just never mapped to
Linux install components.

- stage-release.sh: add gvproxy rows for linux/amd64 and linux/arm64
  (bin/gvproxy, same dest as the darwin row).
- switch: add installed_gvproxy_bin(), one shared definition of where
  an install puts gvproxy (user-local $MINIMAL_BIN / ~/.local/bin,
  else the system-wide /usr/lib/minimal/bin/gvproxy fallback).
- minimald: default the OwnIp switch path through the shared resolver
  instead of the fixed system path, so an installed daemon finds the
  user-local gvproxy. The --gvproxy-bin flag/config still wins.
- minvmd: delegate tiers 2/3 of its resolver to the shared definition
  (its MINVMD_GVPROXY_BIN override tier is unchanged).

The minvmd half of the issue (VM mode on Linux installs) is deferred:
it needs libkrun/libkrunfw and the guest payload staged alongside it,
plus an arm64 minvmd build, all behind the frozen release workflow.

Refs: https://github.com/gominimal/minimal/issues/980
